app/requirements/base.txt
autopep8==1.6.0
boto3==1.20.44
botocore==1.23.44
dj-database-url==0.5.0
Django==4.0.1
django-widget-tweaks==1.4.9
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
docutils==0.18.1
packaging==21.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.5
sendgrid==6.9.5
whitenoise==5.3.0
email-validator==1.1.3
channels==3.0.4
channels-redis==3.3.1
pytest-django==4.5.2
celery==5.2.3
redis==4.1.0
django-redis==5.2.0
flower==1.0.0
locust==2.5.1
pylint==2.12.2
pyfcm==1.5.4
pyotp==2.6.0
sentry-sdk==1.5.4
django-import-export==2.7.1
twilio==7.5.0
djangorestframework==3.13.1
Markdown==3.3.6
MarkupSafe==2.0.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
Faker==11.3.0
selenium==4.1.0
django-health-check==3.16.5
psutil==5.9.0
django-debug-toolbar==3.2.4
pytest==6.2.5
beautifulsoup4==4.10.0
lxml==4.7.1
factory-boy==3.2.1
pytest-celery==0.0.0
